move shadowsocks-libev connectivity tests
