+++ src/electron/sslibev_badvpn_tunnel.ts
@@ -12,16 +12,20 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import * as dgram from 'dgram';
+import * as dns from 'dns';
 import {powerMonitor} from 'electron';
 import {platform} from 'os';
+import * as net from 'net';
 import * as path from 'path';
 import * as process from 'process';
+import * as socks from 'socks';
 
 import {TunnelStatus} from '../www/app/tunnel';
 import * as errors from '../www/model/errors';
 import {ShadowsocksConfig} from '../www/model/shadowsocks';
 
-import {checkUdpForwardingEnabled, isServerReachable, validateServerCredentials} from './connectivity';
+import {isServerReachable} from './connectivity';
 import {ChildProcessHelper} from './process';
 import {RoutingDaemon} from './routing_service';
 import {testTapDevice} from './tap';
@@ -30,9 +34,6 @@ import {VpnTunnel} from './vpn_tunnel';
 
 const isLinux = platform() === 'linux';
 const isWindows = platform() === 'win32';
-
-const PROXY_ADDRESS = '127.0.0.1';
-const PROXY_PORT = 1081;
 
 const TUN2SOCKS_TAP_DEVICE_NAME = isLinux ? 'outline-tun0' : 'outline-tap0';
 const TUN2SOCKS_TAP_DEVICE_IP = '10.0.85.2';
@@ -43,11 +44,19 @@ const TUN2SOCKS_VIRTUAL_ROUTER_NETMASK = '255.255.255.0';
 // ss-local will almost always start, and fast: short timeouts, fast retries.
 const SSLOCAL_CONNECTION_TIMEOUT = 10;
 const SSLOCAL_MAX_ATTEMPTS = 30;
+const SSLOCAL_PROXY_ADDRESS = '127.0.0.1';
+const SSLOCAL_PROXY_PORT = 1081;
 const SSLOCAL_RETRY_INTERVAL_MS = 100;
+
+const CREDENTIALS_TEST_DOMAINS = ['example.com', 'ietf.org', 'wikipedia.org'];
+const DNS_LOOKUP_TIMEOUT_MS = 10000;
+
+const UDP_FORWARDING_TEST_TIMEOUT_MS = 5000;
+const UDP_FORWARDING_TEST_RETRY_INTERVAL_MS = 1000;
 
 async function isSsLocalReachable() {
   await isServerReachable(
-      PROXY_ADDRESS, PROXY_PORT, SSLOCAL_CONNECTION_TIMEOUT, SSLOCAL_MAX_ATTEMPTS,
+      SSLOCAL_PROXY_ADDRESS, SSLOCAL_PROXY_PORT, SSLOCAL_CONNECTION_TIMEOUT, SSLOCAL_MAX_ATTEMPTS,
       SSLOCAL_RETRY_INTERVAL_MS);
 }
 
@@ -61,8 +70,8 @@ async function isSsLocalReachable() {
 //  - silently restart tun2socks when the system is about to suspend (Windows only)
 export class ShadowsocksLibevBadvpnTunnel implements VpnTunnel {
   private readonly routing: RoutingDaemon;
-  private readonly ssLocal = new SsLocal(PROXY_PORT);
-  private readonly tun2socks = new Tun2socks(PROXY_ADDRESS, PROXY_PORT);
+  private readonly ssLocal = new SsLocal(SSLOCAL_PROXY_PORT);
+  private readonly tun2socks = new Tun2socks(SSLOCAL_PROXY_ADDRESS, SSLOCAL_PROXY_PORT);
 
   // Extracted out to an instance variable because in certain situations, notably a change in UDP
   // support, we need to stop and restart tun2socks *without notifying the client* and this allows
@@ -132,10 +141,10 @@ export class ShadowsocksLibevBadvpnTunnel implements VpnTunnel {
     // Don't validate credentials on boot: if the key was revoked, we want the system to stay
     // "connected" so that traffic doesn't leak.
     if (!this.isAutoConnect) {
-      await validateServerCredentials(PROXY_ADDRESS, PROXY_PORT);
+      await validateServerCredentials(SSLOCAL_PROXY_ADDRESS, SSLOCAL_PROXY_PORT);
     }
 
-    this.isUdpEnabled = await checkUdpForwardingEnabled(PROXY_ADDRESS, PROXY_PORT);
+    this.isUdpEnabled = await checkUdpForwardingEnabled(SSLOCAL_PROXY_ADDRESS, SSLOCAL_PROXY_PORT);
     console.log(`UDP support: ${this.isUdpEnabled}`);
     this.tun2socks.start(this.isUdpEnabled);
 
@@ -188,7 +197,7 @@ export class ShadowsocksLibevBadvpnTunnel implements VpnTunnel {
       // Possibly over-cautious, though we have seen occasional failures immediately after network
       // changes: ensure ss-local is reachable first.
       await isSsLocalReachable();
-      if (this.isUdpEnabled === await checkUdpForwardingEnabled(PROXY_ADDRESS, PROXY_PORT)) {
+      if (this.isUdpEnabled === await checkUdpForwardingEnabled(SSLOCAL_PROXY_ADDRESS, SSLOCAL_PROXY_PORT)) {
         return;
       }
     } catch (e) {
@@ -295,4 +304,121 @@ class Tun2socks extends ChildProcessHelper {
 
     this.launch(args);
   }
+}
+
+// Resolves with true iff a response can be received from a semi-randomly-chosen website through the
+// Shadowsocks proxy.
+//
+// This has the same function as ShadowsocksConnectivity.validateServerCredentials in
+// cordova-plugin-outline.
+function validateServerCredentials(proxyAddress: string, proxyIp: number) {
+  return new Promise<void>((fulfill, reject) => {
+    const testDomain =
+        CREDENTIALS_TEST_DOMAINS[Math.floor(Math.random() * CREDENTIALS_TEST_DOMAINS.length)];
+    socks.createConnection(
+        {
+          proxy: {ipaddress: proxyAddress, port: proxyIp, type: 5},
+          target: {host: testDomain, port: 80}
+        },
+        (e, socket) => {
+          if (e) {
+            reject(new errors.InvalidServerCredentials(
+                `could not connect to remote test website: ${e.message}`));
+            return;
+          }
+
+          socket.write(`HEAD / HTTP/1.1\r\nHost: ${testDomain}\r\n\r\n`);
+
+          socket.on('data', (data) => {
+            if (data.toString().startsWith('HTTP/1.1')) {
+              socket.end();
+              fulfill();
+            } else {
+              socket.end();
+              reject(new errors.InvalidServerCredentials(
+                  `unexpected response from remote test website`));
+            }
+          });
+
+          socket.on('close', () => {
+            reject(new errors.InvalidServerCredentials(`could not connect to remote test website`));
+          });
+
+          // Sockets must be resumed before any data will come in, as they are paused right before
+          // this callback is fired.
+          socket.resume();
+        });
+  });
+}
+
+// DNS request to google.com.
+const DNS_REQUEST = Buffer.from([
+  0, 0,                             // [0-1]   query ID
+  1, 0,                             // [2-3]   flags; byte[2] = 1 for recursion desired (RD).
+  0, 1,                             // [4-5]   QDCOUNT (number of queries)
+  0, 0,                             // [6-7]   ANCOUNT (number of answers)
+  0, 0,                             // [8-9]   NSCOUNT (number of name server records)
+  0, 0,                             // [10-11] ARCOUNT (number of additional records)
+  6, 103, 111, 111, 103, 108, 101,  // google
+  3, 99,  111, 109,                 // com
+  0,                                // null terminator of FQDN (root TLD)
+  0, 1,                             // QTYPE, set to A
+  0, 1                              // QCLASS, set to 1 = IN (Internet)
+]);
+
+// Verifies that the remote server has enabled UDP forwarding by sending a DNS request through it.
+function checkUdpForwardingEnabled(proxyAddress: string, proxyIp: number): Promise<boolean> {
+  return new Promise((resolve, reject) => {
+    socks.createConnection(
+        {
+          proxy: {ipaddress: proxyAddress, port: proxyIp, type: 5, command: 'associate'},
+          target: {host: '0.0.0.0', port: 0},  // Specify the actual target once we get a response.
+        },
+        (err, socket, info) => {
+          if (err) {
+            reject(new errors.RemoteUdpForwardingDisabled(`could not connect to local proxy`));
+            return;
+          }
+          const packet = socks.createUDPFrame({host: '1.1.1.1', port: 53}, DNS_REQUEST);
+          const udpSocket = dgram.createSocket('udp4');
+
+          udpSocket.on('error', (e) => {
+            reject(new errors.RemoteUdpForwardingDisabled('UDP socket failure'));
+          });
+
+          udpSocket.on('message', (msg, info) => {
+            stopUdp();
+            resolve(true);
+          });
+
+          // Retry sending the query every second.
+          // TODO: logging here is a bit verbose
+          const intervalId = setInterval(() => {
+            try {
+              udpSocket.send(packet, info.port, info.host, (err) => {
+                if (err) {
+                  console.error(`Failed to send data through UDP: ${err}`);
+                }
+              });
+            } catch (e) {
+              console.error(`Failed to send data through UDP ${e}`);
+            }
+          }, UDP_FORWARDING_TEST_RETRY_INTERVAL_MS);
+
+          const stopUdp = () => {
+            try {
+              clearInterval(intervalId);
+              udpSocket.close();
+            } catch (e) {
+              // Ignore; there may be multiple calls to this function.
+            }
+          };
+
+          // Give up after the timeout elapses.
+          setTimeout(() => {
+            stopUdp();
+            resolve(false);
+          }, UDP_FORWARDING_TEST_TIMEOUT_MS);
+        });
+  });
 }
